Remove sleep from rule engin tests
